# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "## TMF API Reference : TMF 641 - Service Ordering Management\n\n##\
    \ Release : 18.5 - Dec 2018\n\nThe Service Order API provides a standardized mechanism\
    \ for managing Service Order, a type of order which can be used to place an order\
    \ between a customer and a service provider or between a service provider and\
    \ a partner and vice versa.\n\n### Service Order resource\nA service order will\
    \ describe a list of service order items.  A service order item references an\
    \ action on an existing  or  future  service.  By  service  we  designed  Customer\
    \ Facing  Service  (CFS)  as  well  as  Resource Facing Service (RFS).\nFrom a\
    \ component perspective, a service order should be available\n- from a Service\
    \ Orchestration Component (and it could mix CFS and RFS)\n- from an Infrastructure\
    \ Control & Management component (and it would have only RFS)\n\n### Service Order\
    \ API performs the following operations on service order :\n- Retrieval of a service\
    \ order or a collection of service orders depending on filter criteria\n- Partial\
    \ update of a service order (including updating rules)\n- Creation of a service\
    \ order (including default values and creation rules)\n- Deletion of service order\
    \ (for administration purposes)\n- Notification of events on Service order\n\n\
    Copyright © TM Forum 2018. All Rights Reserved"
  version: "4.0.0"
...
host: "localhost:8080"
basePath: "/tmf-api"
...
schemes:
...
consumes:
...
produces:
...
paths:
...
  /listener/serviceOrderCreateNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity ServiceOrderCreateNotification"
      description: "Example of a client listener for receiving the notification ServiceOrderCreateNotification"
      operationId: "listenToServiceOrderCreateNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/ServiceOrderCreateNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/serviceOrderAttributeValueChangeNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity ServiceOrderAttributeValueChangeNotification"
      description: "Example of a client listener for receiving the notification ServiceOrderAttributeValueChangeNotification"
      operationId: "listenToServiceOrderAttributeValueChangeNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/ServiceOrderAttributeValueChangeNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/serviceOrderStateChangeNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity ServiceOrderStateChangeNotification"
      description: "Example of a client listener for receiving the notification ServiceOrderStateChangeNotification"
      operationId: "listenToServiceOrderStateChangeNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/ServiceOrderStateChangeNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/serviceOrderDeleteNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity ServiceOrderDeleteNotification"
      description: "Example of a client listener for receiving the notification ServiceOrderDeleteNotification"
      operationId: "listenToServiceOrderDeleteNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/ServiceOrderDeleteNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
definitions:
...
  EventSubscription:
    type: "object"
    required:
    - "callback"
    - "id"
    properties:
      id:
        type: "string"
        description: "Id of the listener"
      callback:
        type: "string"
        description: "The callback being registered."
      query:
        type: "string"
        description: "additional data to be passed"
    description: "Sets the communication endpoint address the service instance must\
      \ use to deliver notification information"
    example:
      query: "query"
      callback: "callback"
      id: "id"
...
  ServiceOrderCreateNotification:
    type: "object"
    properties:
      eventId:
        type: "string"
        description: "The identifier of the notification"
      eventTime:
        type: "string"
        format: "date-time"
        description: "Time of the event occurrence"
      eventType:
        type: "string"
        description: "The type of the notification"
      fieldPath:
        type: "string"
        description: "The path identifying the object field concerned by this notification"
      resourcePath:
        type: "string"
        description: "The path identifying the resource object concerned by this notification"
      event:
        description: "The event linked to the involved resource object"
        $ref: "#/definitions/ServiceOrderCreateEvent"
    description: "The notification data structure"
    example:
      eventId: "eventId"
      resourcePath: "resourcePath"
      eventTime: "2000-01-23T04:56:07.000+00:00"
      fieldPath: "fieldPath"
      eventType: "eventType"
      event:
        serviceOrder:
          note:
          - date: "2000-01-23T04:56:07.000+00:00"
            system: "system"
            '@baseType': "@baseType"
            author: "author"
            '@type': "@type"
            text: "text"
            '@schemaLocation': "http://example.com/aeiou"
          - date: "2000-01-23T04:56:07.000+00:00"
            system: "system"
            '@baseType': "@baseType"
            author: "author"
            '@type': "@type"
            text: "text"
            '@schemaLocation': "http://example.com/aeiou"
          orderItem:
          - '@baseType': "@baseType"
            service:
              serviceType: "serviceType"
              supportingResource:
              - '@referredType': "@referredType"
                '@baseType': "@baseType"
                '@type': "@type"
                name: "name"
                id: "id"
                href: "href"
                '@schemaLocation': "http://example.com/aeiou"
              - '@referredType': "@referredType"
                '@baseType': "@baseType"
                '@type': "@type"
                name: "name"
                id: "id"
                href: "href"
                '@schemaLocation': "http://example.com/aeiou"
              '@type': "@type"
              serviceSpecification:
                targetServiceSchema:
                  '@baseType': "@baseType"
                  '@type': "@type"
                  '@schemaLocation': "@schemaLocation"
                '@referredType': "@referredType"
                '@baseType': "@baseType"
                '@type': "@type"
                name: "name"
                id: "id"
                href: "href"
                '@schemaLocation': "http://example.com/aeiou"
                version: "version"
              supportingService:
              - '@referredType': "@referredType"
                '@baseType': "@baseType"
                '@type': "@type"
                id: "id"
                href: "href"
                '@schemaLocation': "http://example.com/aeiou"
              - '@referredType': "@referredType"
                '@baseType': "@baseType"
                '@type': "@type"
                id: "id"
                href: "href"
                '@schemaLocation': "http://example.com/aeiou"
              relatedParty:
              - '@referredType': "@referredType"
                role: "role"
                '@baseType': "@baseType"
                '@type': "@type"
                name: "name"
                id: "id"
                href: "href"
                '@schemaLocation': "@schemaLocation"
              - '@referredType': "@referredType"
                role: "role"
                '@baseType': "@baseType"
                '@type': "@type"
                name: "name"
                id: "id"
                href: "href"
                '@schemaLocation': "@schemaLocation"
              serviceRelationship:
              - relationshipType: "relationshipType"
                '@baseType': "@baseType"
                service:
                  '@referredType': "@referredType"
                  '@baseType': "@baseType"
                  '@type': "@type"
                  id: "id"
                  href: "href"
                  '@schemaLocation': "http://example.com/aeiou"
                '@type': "@type"
                '@schemaLocation': "http://example.com/aeiou"
              - relationshipType: "relationshipType"
                '@baseType': "@baseType"
                service:
                  '@referredType': "@referredType"
                  '@baseType': "@baseType"
                  '@type': "@type"
                  id: "id"
                  href: "href"
                  '@schemaLocation': "http://example.com/aeiou"
                '@type': "@type"
                '@schemaLocation': "http://example.com/aeiou"
              '@baseType': "@baseType"
              name: "name"
              id: "id"
              href: "href"
              place:
              - role: "role"
                '@baseType': "@baseType"
                '@type': "@type"
                name: "name"
                id: "id"
                href: "href"
                '@schemaLocation': "http://example.com/aeiou"
              - role: "role"
                '@baseType': "@baseType"
                '@type': "@type"
                name: "name"
                id: "id"
                href: "href"
                '@schemaLocation': "http://example.com/aeiou"
              state: {}
              category: "category"
              '@schemaLocation': "http://example.com/aeiou"
              serviceCharacteristic:
              - '@baseType': "@baseType"
                '@type': "@type"
                valueType: "valueType"
                name: "name"
                '@schemaLocation': "http://example.com/aeiou"
                value: {}
              - '@baseType': "@baseType"
                '@type': "@type"
                valueType: "valueType"
                name: "name"
                '@schemaLocation': "http://example.com/aeiou"
                value: {}
            '@type': "@type"
            action: {}
            appointment:
              '@referredType': "@referredType"
              '@baseType': "@baseType"
              '@type': "@type"
              description: "description"
              id: "id"
              href: "href"
              '@schemaLocation': "http://example.com/aeiou"
            orderItemRelationship:
            - relationshipType: "relationshipType"
              '@baseType': "@baseType"
              '@type': "@type"
              id: "id"
              '@schemaLocation': "http://example.com/aeiou"
            - relationshipType: "relationshipType"
              '@baseType': "@baseType"
              '@type': "@type"
              id: "id"
              '@schemaLocation': "http://example.com/aeiou"
            id: "id"
            state: {}
            '@schemaLocation': "http://example.com/aeiou"
          - '@baseType': "@baseType"
            service:
              serviceType: "serviceType"
              supportingResource:
              - '@referredType': "@referredType"
                '@baseType': "@baseType"
                '@type': "@type"
                name: "name"
                id: "id"
                href: "href"
                '@schemaLocation': "http://example.com/aeiou"
              - '@referredType': "@referredType"
                '@baseType': "@baseType"
                '@type': "@type"
                name: "name"
                id: "id"
                href: "href"
                '@schemaLocation': "http://example.com/aeiou"
              '@type': "@type"
              serviceSpecification:
                targetServiceSchema:
                  '@baseType': "@baseType"
                  '@type': "@type"
                  '@schemaLocation': "@schemaLocation"
                '@referredType': "@referredType"
                '@baseType': "@baseType"
                '@type': "@type"
                name: "name"
                id: "id"
                href: "href"
                '@schemaLocation': "http://example.com/aeiou"
                version: "version"
              supportingService:
              - '@referredType': "@referredType"
                '@baseType': "@baseType"
                '@type': "@type"
                id: "id"
                href: "href"
                '@schemaLocation': "http://example.com/aeiou"
              - '@referredType': "@referredType"
                '@baseType': "@baseType"
                '@type': "@type"
                id: "id"
                href: "href"
                '@schemaLocation': "http://example.com/aeiou"
              relatedParty:
              - '@referredType': "@referredType"
                role: "role"
                '@baseType': "@baseType"
                '@type': "@type"
                name: "name"
                id: "id"
                href: "href"
                '@schemaLocation': "@schemaLocation"
              - '@referredType': "@referredType"
                role: "role"
                '@baseType': "@baseType"
                '@type': "@type"
                name: "name"
                id: "id"
                href: "href"
                '@schemaLocation': "@schemaLocation"
              serviceRelationship:
              - relationshipType: "relationshipType"
                '@baseType': "@baseType"
                service:
                  '@referredType': "@referredType"
                  '@baseType': "@baseType"
                  '@type': "@type"
                  id: "id"
                  href: "href"
                  '@schemaLocation': "http://example.com/aeiou"
                '@type': "@type"
                '@schemaLocation': "http://example.com/aeiou"
              - relationshipType: "relationshipType"
                '@baseType': "@baseType"
                service:
                  '@referredType': "@referredType"
                  '@baseType': "@baseType"
                  '@type': "@type"
                  id: "id"
                  href: "href"
                  '@schemaLocation': "http://example.com/aeiou"
                '@type': "@type"
                '@schemaLocation': "http://example.com/aeiou"
              '@baseType': "@baseType"
              name: "name"
              id: "id"
              href: "href"
              place:
              - role: "role"
                '@baseType': "@baseType"
                '@type': "@type"
                name: "name"
                id: "id"
                href: "href"
                '@schemaLocation': "http://example.com/aeiou"
              - role: "role"
                '@baseType': "@baseType"
                '@type': "@type"
                name: "name"
                id: "id"
                href: "href"
                '@schemaLocation': "http://example.com/aeiou"
              state: {}
              category: "category"
              '@schemaLocation': "http://example.com/aeiou"
              serviceCharacteristic:
              - '@baseType': "@baseType"
                '@type': "@type"
                valueType: "valueType"
                name: "name"
                '@schemaLocation': "http://example.com/aeiou"
                value: {}
              - '@baseType': "@baseType"
                '@type': "@type"
                valueType: "valueType"
                name: "name"
                '@schemaLocation': "http://example.com/aeiou"
                value: {}
            '@type': "@type"
            action: {}
            appointment:
              '@referredType': "@referredType"
              '@baseType': "@baseType"
              '@type': "@type"
              description: "description"
              id: "id"
              href: "href"
              '@schemaLocation': "http://example.com/aeiou"
            orderItemRelationship:
            - relationshipType: "relationshipType"
              '@baseType': "@baseType"
              '@type': "@type"
              id: "id"
              '@schemaLocation': "http://example.com/aeiou"
            - relationshipType: "relationshipType"
              '@baseType': "@baseType"
              '@type': "@type"
              id: "id"
              '@schemaLocation': "http://example.com/aeiou"
            id: "id"
            state: {}
            '@schemaLocation': "http://example.com/aeiou"
          '@type': "@type"
          description: "description"
          externalId: "externalId"
          priority: "priority"
          relatedParty:
          - '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          - '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          requestedStartDate: "2000-01-23T04:56:07.000+00:00"
          '@baseType': "@baseType"
          notificationContact: "notificationContact"
          completionDate: "2000-01-23T04:56:07.000+00:00"
          expectedCompletionDate: "2000-01-23T04:56:07.000+00:00"
          id: "id"
          href: "href"
          category: "category"
          orderRelationship:
          - '@referredType': "@referredType"
            relationshipType: "relationshipType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - '@referredType': "@referredType"
            relationshipType: "relationshipType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          '@schemaLocation': "http://example.com/aeiou"
          orderDate: "2000-01-23T04:56:07.000+00:00"
          startDate: "2000-01-23T04:56:07.000+00:00"
          requestedCompletionDate: "2000-01-23T04:56:07.000+00:00"
...
  Error:
    type: "object"
    required:
    - "code"
    - "reason"
    properties:
      code:
        type: "integer"
        description: "Application relevant detail, defined in the API or a common\
          \ list."
      reason:
        type: "string"
        description: "Explanation of the reason for the error which can be shown to\
          \ a client user."
      message:
        type: "string"
        description: "More details and corrective actions related to the error which\
          \ can be shown to a client user."
      status:
        type: "integer"
        description: "HTTP Error code extension"
      referenceError:
        type: "string"
        format: "uri"
        description: "URI of documentation describing the error."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class."
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name."
    description: "Used when an API throws an Error, typically with a HTTP error response-code\
      \ (3xx, 4xx, 5xx)"